requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.10.7
greenlet==2.0.2
gunicorn==20.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
platformdirs==3.2.0
psycopg2==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
rest-framework-simplejwt==0.0.2
six==1.16.0
SQLAlchemy==1.4.17
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
virtualenv==20.21.0
